Crate dock_runtime

Dock testnet runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use poa;

pub use price_feed;

pub use token_migration;

Modules

anchor

Minimal proof of existence registry.

api

attest

This module allows DIDs to publically attests to arbirary (and arbitrarily large) RDF claimgraphs. These attestations are not stored on-chain; rather, the attester chooses a storage method by specifying an Iri.

blob

Generic immutable single-owner storage.

did

master

Simulates a multisig root account.

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

revoke

weight_to_fee

Structs

AdjustmentVariable

BaseFilter

BasicDeposit

BlockGasLimit

BlockHashCount

BondingDuration

Bonding duration is in number of era

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

CandidacyBond

CooloffPeriod

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DataDepositPerByte

DealWithFees

Deals with transaction fees

DesiredMembers

DesiredRunnersUp

DisabledValidatorsThreshold

DockChainId

ElectionLookahead

ElectionsPhragmenModuleId

EnactmentPeriod

EpochDuration

EthereumFindAuthor

ExistentialDeposit

ExpectedBlockTime

Fallback

FastTrackVotingPeriod

FieldDeposit

GasPrice

Fixed gas price

GasWeightMap

GenesisConfig

ImOnlineUnsignedPriority

InstantAllowed

LaunchPeriod

MaxAdditionalFields

MaxBlobSize

MaxIterations

MaxLocks

MaxNominatorRewardedPerValidator

A validator will only have 256 nominators, so max (rewarded) nominators is 256*50 = 12800. This number is smaller than current token holders (20K) but not all holder participate. Keeping it small to batch payout extrinsics

MaxProposals

MaxRegistrars

MaxScheduledPerBlock

MaxSubAccounts

MaxVotes

MaximumBlockWeight

We allow for 1 seconds of compute with a 3 second average block time.

MaximumReasonLength

MinDockFiatRate

Price of Dock/USD pair as 10th of cent. Value of 10 means 1 cent

MinSolutionScoreBump

MinerMaxIterations

MinerMaxWeight

MinimumDeposit

1000 tokens

MinimumMultiplier

MinimumPeriod

MultiPhaseUnsignedPriority

OffchainSolutionWeightLimit

OffencesWeightSoftLimit

Origin

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageByteDeposit

0.1 token

ProposalBond

ProposalBondMinimum

ReportLongevity

Specifies the number of blocks for which the equivocation is valid.

RewardCurve

RewardDecayPct

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SessionsPerEra

SignedPhase

SlashDeferDuration

SolutionImprovementThreshold

SpendPeriod

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

StorageWeight

SubAccountDeposit

TargetBlockFullness

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

.01 token

TransactionConverter

TreasuryModuleId

Matches treasury account created during PoA.

TreasuryRewardsPct

U64CurrencyToVote

UncleGenerations

UnsignedPhase

Version

VestingDuration

Vesting duration in number of blocks. Duration is 183 days and block time is 3 sec. (183 * 24 * 3600) / 3 = 5270400

VestingMilestones

Number of vesting milestones

VotingBondBase

VotingBondFactor

VotingPeriod

Enums

Call

Event

OriginCaller

StakerStatus

Indicates the initial status of the staker.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 10% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 10%.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

DAYS

DOCK

The token has 6 decimal places

HOURS

MAXIMUM_BLOCK_WEIGHT

We allow for 1 second of compute with a 3 second average block time.

MILLISECS_PER_BLOCK

MINUTES

PRIMARY_PROBABILITY

VERSION

This runtime version. Whe compiling with feature “mainnet”, spec_name will be “dock-main-runtime”, when compiling with feature “testnet”, spec_name will be “dock-test-runtime”, when not using either of those, spec_name will be “dock-dev-runtime”

WASM_BINARY

WEIGHT_PER_GAS

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address

The address format for describing accounts.

AllModules

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem

All pallets included in the runtime as a nested tuple of types.

Anchor

Attest

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balance

Balance of an account.

Balances

BalancesConfig

BlobStore

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

Bounties

CheckedExtrinsic

Extrinsic type that has already been checked.

Council

CouncilConfig

DIDModule

DIDModuleConfig

Democracy

EVM

EVMConfig

ElectionProviderMultiPhase

Elections

ElectionsConfig

Ethereum

EthereumConfig

FiatFilterModule

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Historical

Identity

ImOnline

ImOnlineConfig

Index

Index of a transaction in the chain.

Master

MasterConfig

MigrationModule

Moment

Type used for expressing timestamp.

Offences

PoAModule

PoAModuleConfig

PriceFeedModule

Revoke

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedPayload

The payload being signed in transactions.

Staking

StakingConfig

StakingRewards

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeMembership

TechnicalCommitteeMembershipConfig

Timestamp

Tips

TransactionPayment

Treasury

TreasuryConfig

Utility