Module dock_runtime::anchor

Minimal proof of existence registry.

Anchors are hashed once before being added to storage. To check whether an anchor exists query the “Anchors” map for the hash of the anchor. If a corresponding value exists, then the anchor exists and the value represents the block number when it was first published.

Structs

Module

Enums

Call

Dispatchable calls.

Error

RawEvent

Events for this module.

Traits

Trait

_

Provides version information about a pallet.

Type Definitions

Event

RawEvent specialized for the configuration Config