Module dock_runtime::common

Modules

authorization

keys

limits

policy

signatures

signed_action

state_change

storage_version

types

Structs

Authorization

Successfully authorized signer along with its key.

CombineTypesAndLimits

Combines two different types - first implementing Types and second implementing Limits to form a single type implementing both.

MultiSignedAction

An action signed by multiple signers with their corresponding nonces.

PolicyExecutor

DID performing an action according to the policies.

Secp256k1Keypair

SigTypes

SignatureWithNonce

DIDs signature along with the nonce.

SignedActionWithNonce

Enums

CurveType

DidMethodKeySigValue

An abstraction for a signature.

IntermediateError

Either DispatchError or DispatchErrorWithPostInfo.

Policy

Authorization logic containing rules to modify some data entity.

PolicyValidationError

An error occurred during Policy validation.

PublicKey

An abstraction for a public key. Abstracts the type and value of the public key where the value is a byte array

SigValue

An abstraction for a signature.

StateChange

Any state change that needs to be signed is first wrapped in this enum and then its serialized. This is done to make it unambiguous which command was intended as the SCALE codec’s not self describing. The enum variants are supposed to take care of replay protection by having a nonce or something else. A better approach would have been to make StateChange aware of nonce or nonces. There can be multiple nonces attached with a payload a multiple DIDs may take part in an action and they will have their own nonce. However this change will be a major disruption for now. Never change the order of variants in this enum

StorageVersion

Defines version of the storage being used.

VerificationError

Constants

ED25519_WEIGHT

SECP256K1_WEIGHT

SR25519_WEIGHT

Traits

AuthorizeSignedAction

Signature that can authorize a signed action.

AuthorizeTarget

Authorizes action performed by Self over supplied target using given key.

ForSigType

GetKey

Attempts to retrieve a key associated with the given entity.

Limits

All size limits for the dock-core encodable data structures.

Signature

Signature entity.

Size

ToStateChange

Converts the given entity to the state change.

Types

Defines associated types used by dock-core.

TypesAndLimits

All associated types and size limits for the encodable data structures used by the dock-core.

Functions

get_secp256k1_keypair

get_secp256k1_keypair_struct

sign_with_secp256k1