Trait dock_runtime::common::authorization::AuthorizeTarget

pub trait AuthorizeTarget<T, Target, Key>where
    Target: Associated<T>,{
    // Provided method
    fn ensure_authorizes_target<A>(
        &self,
        _: &Key,
        _: &A,
        _: Option<&<Target as Associated<T>>::Value>
    ) -> Result<(), DispatchError>
       where A: Action<Target = Target> { ... }
}

Authorizes action performed by Self over supplied target using given key.

Provided Methods

fn ensure_authorizes_target<A>( &self, _: &Key, _: &A, _: Option<&<Target as Associated<T>>::Value> ) -> Result<(), DispatchError>where A: Action<Target = Target>,

Self can perform supplied action over target using the provided key.