Trait dock_runtime::common::limits::Limits

pub trait Limits: Clone + Eq {

    type MaxAccumulatorLabelSize: Size;
    type MaxAccumulatorParamsSize: Size;
    type MaxAccumulatorPublicKeySize: Size;
    type MaxAccumulatorAccumulatedSize: Size;
    type MaxDidDocRefSize: Size;
    type MaxDidServiceEndpointIdSize: Size;
    type MaxDidServiceEndpointOrigins: Size;
    type MaxDidServiceEndpointOriginSize: Size;
    type MaxStatusListCredentialSize: Size;
    type MinStatusListCredentialSize: Size;
    type MaxIriSize: Size;
    type MaxBlobSize: Size;
    type MaxOffchainParamsLabelSize: Size;
    type MaxOffchainParamsBytesSize: Size;
    type MaxBBSPublicKeySize: Size;
    type MaxBBSPlusPublicKeySize: Size;
    type MaxBBDT16PublicKeySize: Size;
    type MaxPSPublicKeySize: Size;
    type MaxMasterMembers: Size;
    type MaxPolicyControllers: Size;
    type MaxIssuerPriceCurrencySymbolSize: Size;
    type MaxIssuersPerSchema: Size;
    type MaxVerifiersPerSchema: Size;
    type MaxIssuerPriceCurrencies: Size;
    type MaxTrustRegistryNameSize: Size;
    type MaxConvenerRegistries: Size;
    type MaxDelegatedIssuers: Size;
    type MaxSchemasPerIssuer: Size;
    type MaxSchemasPerVerifier: Size;
    type MaxRegistriesPerIssuer: Size;
    type MaxRegistriesPerVerifier: Size;
    type MaxTrustRegistryGovFrameworkSize: Size;
    type MaxSchemasPerRegistry: Size;
    type MaxParticipantsPerRegistry: Size;
    type MaxRegistryParticipantOrgNameSize: Size;
    type MaxRegistryParticipantDescriptionSize: Size;
    type MaxRegistryParticipantLogoSize: Size;

}

All size limits for the dock-core encodable data structures.

Required Associated Types

type MaxAccumulatorLabelSize: Size

Maximum size of the label

type MaxAccumulatorParamsSize: Size

Maximum byte size of the parameters. This depends only on the chosen elliptic curve.

type MaxAccumulatorPublicKeySize: Size

Maximum byte size of the public key. This depends only on the chosen elliptic curve.

type MaxAccumulatorAccumulatedSize: Size

Maximum byte size of the accumulated value which is just one group element (not the number of members)

type MaxDidDocRefSize: Size

Maximum byte size of reference to off-chain DID Doc.

type MaxDidServiceEndpointIdSize: Size

Maximum byte size of service endpoint’s id field

type MaxDidServiceEndpointOrigins: Size

Maximum number of service endpoint’s origin

type MaxDidServiceEndpointOriginSize: Size

Maximum byte size of service endpoint’s origin

type MaxStatusListCredentialSize: Size

StatusListCredentials with size larger than this won’t be accepted.

type MinStatusListCredentialSize: Size

StatusListCredentials with size less than this won’t be accepted.

type MaxIriSize: Size

Max byte size of the Iri.

type MaxBlobSize: Size

Max byte size of the Blob.

type MaxOffchainParamsLabelSize: Size

Max byte size of the offchain params label.

type MaxOffchainParamsBytesSize: Size

Max byte size of the offchain params bytes.

type MaxBBSPublicKeySize: Size

Maximum byte size of the BBS (fixed size) public key. This depends only on the chosen elliptic curve.

type MaxBBSPlusPublicKeySize: Size

Maximum byte size of the BBS+ (fixed size) public key. This depends only on the chosen elliptic curve.

type MaxBBDT16PublicKeySize: Size

Maximum byte size of the BBDT16 (fixed size) public key. This depends only on the chosen elliptic curve.

type MaxPSPublicKeySize: Size

Maximum byte size of the PS public key. This depends on the chosen elliptic curve and the number of messages that can be signed.

type MaxMasterMembers: Size

Max amount of master members per a single Membership.

type MaxPolicyControllers: Size

Max amount of the controller DIDs per a single Policy.

type MaxIssuerPriceCurrencySymbolSize: Size

type MaxIssuersPerSchema: Size

Max no of issuers that can issue for a particular schema

type MaxVerifiersPerSchema: Size

Max no of verifiers that can verify (credential) for a particular schema

type MaxIssuerPriceCurrencies: Size

Max no of currencies an issuer can set the price in for a particular schema

type MaxTrustRegistryNameSize: Size

Max size of the trust registry name.

type MaxConvenerRegistries: Size

Max no of the trust registries per a single convener.

type MaxDelegatedIssuers: Size

Max no of delegated issuers a particular issuer can have

type MaxSchemasPerIssuer: Size

Max no of schemas that a particular issuer can issue for

type MaxSchemasPerVerifier: Size

Max no of schemas that a particular verifier can verify (credential) for

type MaxRegistriesPerIssuer: Size

Max no of registries per a single issuer.

type MaxRegistriesPerVerifier: Size

Max no of registries per a single verifier.

type MaxTrustRegistryGovFrameworkSize: Size

Max government framework size for the trust registry.

type MaxSchemasPerRegistry: Size

Max no of schemas per registry.

type MaxParticipantsPerRegistry: Size

Max no of participants per a single registry.

type MaxRegistryParticipantOrgNameSize: Size

Max size of the TrustRegistry participant’s org name.

type MaxRegistryParticipantDescriptionSize: Size

Max size of the TrustRegistry participant’s description.

type MaxRegistryParticipantLogoSize: Size

Max size of the TrustRegistry participant’s logo.