Enum dock_runtime::did::Call

pub enum Call<T>where
    T: Config,{

    new_offchain {
        did: Did,
        did_doc_ref: OffChainDidDocRef<T>,
    },
    set_offchain_did_doc_ref {
        did: Did,
        did_doc_ref: OffChainDidDocRef<T>,
    },
    remove_offchain_did {
        did: Did,
    },
    new_onchain {
        did: Did,
        keys: Vec<UncheckedDidKey, Global>,
        controllers: BTreeSet<Controller, Global>,
    },
    add_keys {
        keys: AddKeys<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    remove_keys {
        keys: RemoveKeys<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    add_controllers {
        controllers: AddControllers<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    remove_controllers {
        controllers: RemoveControllers<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    add_service_endpoint {
        service_endpoint: AddServiceEndpoint<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    remove_service_endpoint {
        service_endpoint: RemoveServiceEndpoint<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    remove_onchain_did {
        removal: DidRemoval<T>,
        sig: DidOrDidMethodKeySignature<Controller>,
    },
    new_did_method_key {
        did_key: DidMethodKey,
    },
    noop {
        s: StateChange<'static, T>,
        d: AggregatedDidDetailsResponse<T>,
        qi: QueryTrustRegistryBy,
        qy: QueryTrustRegistriesBy,
        a: AggregatedTrustRegistrySchemaMetadata<T>,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

new_offchain

Fields

did: Did

did_doc_ref: OffChainDidDocRef<T>

Creates a new offchain DID (Decentralized Identifier) entry.

This function is used to create a new offchain DID entry by providing a reference to an offchain DID document.

Parameters

origin: The origin of the call, which determines who is making the request.

did: The decentralized identifier (DID) that uniquely identifies the entity.

did_doc_ref: The new reference to the offchain DID document. It can be one of the following:

CID: A Content Identifier as per multiformats/cid.

URL: A URL pointing to the DID document.

Custom: A custom encoding of the reference.

set_offchain_did_doc_ref

Fields

did: Did

did_doc_ref: OffChainDidDocRef<T>

Updates the offchain DID document reference for an existing DID.

This function is used to set or update the reference to the offchain DID document for a given DID. The offchain DID document reference can be one of the following types: CID, URL, or Custom.

Parameters

origin: The origin of the call, which determines who is making the request and their permissions.

did: The decentralized identifier (DID) that uniquely identifies the entity whose DID document reference is being updated.

did_doc_ref: The new reference to the offchain DID document. It can be one of the following:

CID: A Content Identifier as per multiformats/cid.

URL: A URL pointing to the DID document.

Custom: A custom encoding of the reference.

remove_offchain_did

Fields

did: Did

Removes an existing offchain DID entry.

This function is used to remove an offchain DID entry from the system. This operation deletes the DID and its associated offchain DID document reference.

Parameters

origin: The origin of the call, which determines who is making the request and their permissions.

did: The decentralized identifier (DID) that uniquely identifies the entity to be removed.

new_onchain

Fields

did: Did

keys: Vec<UncheckedDidKey, Global>

controllers: BTreeSet<Controller, Global>

Create new DID. At least 1 control key or 1 controller must be provided. If any supplied key has an empty ver_rel, then it will use all verification relationships available for its key type.

add_keys

Fields

keys: AddKeys<T>

sig: DidOrDidMethodKeySignature<Controller>

Add more keys from DID doc. Does not check if the key was already added.

remove_keys

Fields

keys: RemoveKeys<T>

sig: DidOrDidMethodKeySignature<Controller>

Remove keys from DID doc. This is an atomic operation meaning that it will either remove all keys or do nothing. Note that removing all keys might make DID unusable.

add_controllers

Fields

controllers: AddControllers<T>

sig: DidOrDidMethodKeySignature<Controller>

Add new controllers to the signer DID. Does not require provided controllers to

have any key

exist on- or off-chain

remove_controllers

Fields

controllers: RemoveControllers<T>

sig: DidOrDidMethodKeySignature<Controller>

Remove controllers from the signer DID. This is an atomic operation meaning that it will either remove all keys or do nothing. Note that removing all controllers might make DID unusable.

add_service_endpoint

Fields

service_endpoint: AddServiceEndpoint<T>

sig: DidOrDidMethodKeySignature<Controller>

Add a single service endpoint to the signer DID.

remove_service_endpoint

Fields

service_endpoint: RemoveServiceEndpoint<T>

sig: DidOrDidMethodKeySignature<Controller>

Remove a single service endpoint.

remove_onchain_did

Fields

removal: DidRemoval<T>

sig: DidOrDidMethodKeySignature<Controller>

Remove the on-chain DID along with its keys, controllers, service endpoints and BBS+ keys. Other DID-controlled entities won’t be removed. However, the authorization logic ensures that once a DID is removed, it loses its ability to control any DID.

new_did_method_key

Fields

did_key: DidMethodKey

Adds an on-chain state storing the nonce for the provided DID method key. After this state is set, this DID method key will be able to submit a DID transaction.

noop

Fields

s: StateChange<'static, T>

d: AggregatedDidDetailsResponse<T>

qi: QueryTrustRegistryBy

qy: QueryTrustRegistriesBy

a: AggregatedTrustRegistrySchemaMetadata<T>

Adds StateChange and AggregatedDidDetailsResponse to the metadata.

Implementations

impl<T> Call<T>where T: Config,

pub fn new_call_variant_new_offchain( did: Did, did_doc_ref: OffChainDidDocRef<T> ) -> Call<T>

Create a call with the variant new_offchain.

pub fn new_call_variant_set_offchain_did_doc_ref( did: Did, did_doc_ref: OffChainDidDocRef<T> ) -> Call<T>

Create a call with the variant set_offchain_did_doc_ref.

pub fn new_call_variant_remove_offchain_did(did: Did) -> Call<T>

Create a call with the variant remove_offchain_did.

pub fn new_call_variant_new_onchain( did: Did, keys: Vec<UncheckedDidKey, Global>, controllers: BTreeSet<Controller, Global> ) -> Call<T>

Create a call with the variant new_onchain.

pub fn new_call_variant_add_keys( keys: AddKeys<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant add_keys.

pub fn new_call_variant_remove_keys( keys: RemoveKeys<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant remove_keys.

pub fn new_call_variant_add_controllers( controllers: AddControllers<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant add_controllers.

pub fn new_call_variant_remove_controllers( controllers: RemoveControllers<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant remove_controllers.

pub fn new_call_variant_add_service_endpoint( service_endpoint: AddServiceEndpoint<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant add_service_endpoint.

pub fn new_call_variant_remove_service_endpoint( service_endpoint: RemoveServiceEndpoint<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant remove_service_endpoint.

pub fn new_call_variant_remove_onchain_did( removal: DidRemoval<T>, sig: DidOrDidMethodKeySignature<Controller> ) -> Call<T>

Create a call with the variant remove_onchain_did.

pub fn new_call_variant_new_did_method_key(did_key: DidMethodKey) -> Call<T>

Create a call with the variant new_did_method_key.

pub fn new_call_variant_noop( s: StateChange<'static, T>, d: AggregatedDidDetailsResponse<T>, qi: QueryTrustRegistryBy, qy: QueryTrustRegistriesBy, a: AggregatedTrustRegistrySchemaMetadata<T> ) -> Call<T>

Create a call with the variant noop.