Struct dock_runtime::did::Module

pub struct Module<T>(_)
 where
    T: Trait;

Implementations

impl<T> Module<T> where
T: Trait,

pub fn get_key_detail(
did: &[u8; 32]
) -> Result<(KeyDetail, <T as Config>::BlockNumber), DispatchError>

Get the key detail and the block number of last modification of the given DID. It assumes that the DID has only 1 public key which is true for now but will change later. This function will then be modified to indicate which key(s) of the DID should be used. If DID is not registered an error is raised.

pub fn verify_sig_from_did(
signature: &DidSignature,
message: &[u8],
did: &[u8; 32]
) -> Result<bool, DispatchError>

Verify given signature on the given message with the given DID’s only public key. It assumes that the DID has only 1 public key which is true for now but will change later. This function will then be modified to indicate which key(s) of the DID should be used. If DID is not registered an error is raised. This function is intended to be used by other modules as well to check the signature from a DID.

pub fn verify_sig_with_public_key(
signature: &DidSignature,
message: &[u8],
public_key: &PublicKey
) -> Result<bool, DispatchError>

Verify given signature on the given message with given public key

impl<T> Module<T> where
T: 'static + Trait,

pub fn did<K>(key: K) -> Option<(KeyDetail, <T as Config>::BlockNumber)> where
K: EncodeLike<[u8; 32]>,

impl<T> Module<T> where
T: Trait,

Can also be called using Call.

pub fn new(
origin: <T as Config>::Origin,
did: [u8; 32],
detail: KeyDetail
) -> Result<(), DispatchError>

Create a new DID. did is the new DID to create. The method will fail if did is already registered. detail is the details of the key like its type, controller and value. The controller DID is not required to exist in the state.

NOTE: Calling this function will bypass origin filters.

pub fn update_key(
origin: <T as Config>::Origin,
key_update: KeyUpdate,
signature: DidSignature
) -> Result<(), DispatchError>

Sets the single publicKey (and possibly its controller) stored in this DID.

key_update specifies which DID’s key needs to be updated. The new controller DID is not required to exist in the state. signature is the signature on a serialized StateChange that wraps the KeyUpdate struct

During execution this function checks for a signature over StateChange and verifies the given signature with the stored key.

This call requires a signature verification and the cost of verification varies by type of signature

NOTE: Calling this function will bypass origin filters.

pub fn remove(
origin: <T as Config>::Origin,
to_remove: DidRemoval,
signature: DidSignature
) -> Result<(), DispatchError>

Deletes a DID from chain storage. Once the DID is deleted, anyone can call new to claim it for their own.

to_remove contains the DID to be removed signature is the signature on a serialized StateChange that wraps the DidRemoval struct

During execution this function checks for a signature over StateChange and verifies the given signature with the stored key.

This call requires a signature verification and the cost of verification varies by type of signature

NOTE: Calling this function will bypass origin filters.