Module dock_runtime::offchain_signatures

Module to store offchain signature keys and parameters for different signature schemes. Currently, can be either BBS, BBS+, Pointcheval-Sanders or BBDT16. For BBS, BBS+ and Pointcheval-Sanders, the public key is in group G2 but for BBDT16, it’s in group G1 and used to verify the proof of validity of MAC (and related proofs) but not the MAC itself.

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

AddOffchainSignatureParams

AddOffchainSignaturePublicKey

BBDT16Parameters

Signature parameters for the BBDT16 signature scheme.

BBDT16PublicKey

Public key for the BBDT16 signature scheme.

BBSParameters

Signature parameters for the BBS signature scheme.

BBSPlusParameters

Signature parameters for the BBS+ signature scheme.

BBSPlusPublicKey

Public key for the BBS+ signature scheme.

BBSPublicKey

Public key for the BBS signature scheme.

GenesisConfig

Can be used to configure the genesis state of this pallet.

PSParameters

Signature parameters for the PS signature scheme.

PSPublicKey

Public key for the PS signature scheme.

Pallet

The pallet implementing the on-chain logic.

RemoveOffchainSignatureParams

RemoveOffchainSignaturePublicKey

SignatureParamsOwner

DID owner of the signature parameters.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

OffchainPublicKey

Public key for different signature schemes. Currently, can be either BBS, BBS+, Pointcheval-Sanders or BBDT16.

OffchainSignatureParams

Signature parameters. Currently can be either BBS, BBS+ or Pointcheval-Sanders.

Traits

Config

The module’s configuration trait.

Type Definitions

BBDT16PublicKeyWithParams

BBSPlusPublicKeyWithParams

BBSPublicKeyWithParams

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PSPublicKeyWithParams

ParamsCounter

On adding new params, corresponding counter is increased by 1 but the counters don’t decrease on removal.

ParticipantId

Identifier of the participant used in the threshold issuance.

PublicKeys

Public keys are stored as key value (did, counter) -> public key

SignatureParams

Signature parameters are stored as key value (did, counter) -> signature parameters

SignatureParamsStorageKey

SignaturePublicKeyStorageKey

Version