Module dock_runtime::trust_registry

Dock Trust Registry.

Modules

actions

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

query

types

Structs

AggregatedIssuerInfo

AggregatedTrustRegistrySchemaMetadata

TrustRegistry schema metadata.

ChangeParticipantsRaw

Convener

Trust registry Convener’s DID.

ConvenerOrIssuerOrVerifier

Trust registry Convener/Issuer/Verifier’s DID.

DelegatedIssuerSchemas

Set of schemas corresponding to a delegated Issuer

DelegatedIssuers

Delegated Issuers.

DelegatedSchemaCounter

InitOrUpdateTrustRegistry

Issuer

Trust registry Issuer’s DID.

IssuerAndVerifier

Both an Issuer and a Verifier.

IssuerOrVerifier

Maybe an Issuer or a Verifier but definitely not a Convener.

IssuerSchemas

Set of schemas corresponding to an Issuer.

IssuerTrustRegistries

Set of trust registries corresponding to a issuer

IssuersOrVerifiers

IssuersWith

A map from Issuer to some value.

MapToPairs

A wrapper for the iterator that converts an iterator of values to an iterator of pairs obtained using supplied functions.

Pallet

The pallet implementing the on-chain logic.

QueryTrustRegistriesBy

Specifies arguments to retrieve trust registries informations by.

QueryTrustRegistryBy

Specifies arguments to retrieve trust registry informations by.

Schemas

SetParticipantInformationRaw

SetSchemasMetadata

StorageAccesses

Number of times storage entities were accessed.

SuspendIssuers

TrustRegistryId

Unique identifier for the TrustRegistry.

TrustRegistryIdForParticipants

TrustRegistryIdSet

TrustRegistryInfo

TrustRegistryIssuerConfiguration

TrustRegistrySchemaId

Unique identifier for the TrustRegistry.

TrustRegistrySchemaMetadata

TrustRegistry schema metadata.

TrustRegistrySchemaVerifiers

Schema Verifiers.

TrustRegistryStoredParticipantInformation

Information about the TrustRegistry participant.

TrustRegistryStoredParticipants

Set of TrustRegistry participants.

TrustRegistryStoredSchemas

Set of schemas that belong to the TrustRegistry

UnboundedDelegatedIssuers

Unbounded delegated Issuers.

UnboundedIssuersWith

An unbounded map from Issuer to some value.

UnboundedSchemas

UnboundedTrustRegistryParticipantInformation

Information about the TrustRegistry participant.

UnboundedTrustRegistrySchemaMetadata

Unbounded TrustRegistry schema metadata.

UnboundedTrustRegistrySchemaVerifiers

Schema Verifiers.

UnboundedVerificationPrices

Prices of verifying a credential corresponding to the specific schema metadata per different currencies.

UnsuspendIssuers

UpdateDelegatedIssuers

VerificationPrice

Price to verify a credential. Lowest denomination should be used.

VerificationPrices

Prices of verifying a credential corresponding to the specific schema metadata per different currencies.

Verifier

Trust registry Verifier’s DID.

VerifierSchemas

Set of schemas corresponding to a verifier

VerifierTrustRegistries

Set of trust registries corresponding to a verifier

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Error for the TrustRegistry module.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

HasSchemasAndRegistries

An entity that has references to schemas and registries.

IntoModuleError

Type Definitions

AggregatedTrustRegistrySchemaIssuers

ChangeParticipants

ConvenerTrustRegistries

Stores TrustRegistrys owned by conveners as a mapping of the form convener_id -> Set<registry_id>

DelegatedUpdate

IssuersTrustRegistries

Stores a set of Issuers Trust Registries.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

MultiSchemaIdUpdate

RegistryIdUpdate

SchemaIdUpdate

SetParticipantInformation

TrustRegistriesInfo

Stores TrustRegistrys information: Convener, name, etc.

TrustRegistriesParticipants

Trust Registry participants. Mapping of TrustRegistryId -> set of participants (Verifiers and Issuers).

TrustRegistriesStoredSchemas

Schema ids corresponding to trust registries. Mapping of TrustRegistryId -> set of schema ids.

TrustRegistryDelegatedIssuerSchemas

Stores Trust Registry’s delegated Issuers schemas.

TrustRegistryIssuerConfigurations

Stores Trust Registry’s Issuers configurations.

TrustRegistryIssuerSchemas

Stores Trust Registry’s Issuers schemas.

TrustRegistryParticipantsInformation

Trust Registry participants. Mapping of TrustRegistryId -> Issuer -> trust registry participant information.

TrustRegistrySchemaIssuers

Schema Issuers (Issuer => verification prices).

TrustRegistrySchemasMetadata

Schema metadata stored in all trust registries. Mapping of the form (schema_id, registry_id) -> schema_metadata

TrustRegistryVerifierSchemas

Stores Trust Registry’s Verifiers schemas.

UnboundedDelegatedIssuersUpdate

UnboundedRegistrySchemaIssuers

UnboundedTrustRegistryParticipantsUpdate

VerifiersTrustRegistries

Stores a set of Verifiers Trust Registries.