Crate dock_runtime

Dock blockchain runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use dock_poa;

pub use dock_price_feed;

pub use dock_token_migration;

Modules

accumulator

anchor

Minimal proof of existence registry.

api

attest

This module allows DIDs to publically attests to arbirary (and arbitrarily large) RDF claimgraphs. These attestations are not stored on-chain; rather, the attester chooses a storage method by specifying an Iri.

blob

Generic immutable single-owner storage.

common

did

master

Simulates a multisig root account.

offchain_signatures

Module to store offchain signature keys and parameters for different signature schemes. Currently, can be either BBS, BBS+, Pointcheval-Sanders or BBDT16. For BBS, BBS+ and Pointcheval-Sanders, the public key is in group G2 but for BBDT16, it’s in group G1 and used to verify the proof of validity of MAC (and related proofs) but not the MAC itself.

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

precompiles

revoke

status_list_credential

Storage for status list-related verifiable credentials:

trust_registry

Dock Trust Registry.

weight_to_fee

Structs

AdjustmentVariable

BBSPlusPublicKey

Public key for the BBS+ signature scheme.

BaseFeeThreshold

Denotes when the system should decrease/increase the base fee based on how much weight was used by the last block. target is the ideal congestion of the network where the base fee should remain unchanged. Under normal circumstances the target should be 50%. If we go below the target, the base fee is linearly decreased by the Elasticity delta of lower~target. If we go above the target, the base fee is linearly increased by the Elasticity delta of upper~target. The base fee is fully increased (default 12.5%) if the block is upper full (default 100%). The base fee is fully decreased (default 12.5%) if the block is lower empty (default 0%).

BasicDeposit

BenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

BetterUnsignedThreshold

BlockGasLimit

BlockHashCount

BondingDuration

Bonding duration is in number of era

BountyCuratorDepositMax

BountyCuratorDepositMin

BountyCuratorDepositMultiplier

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

We are fixed supply token, we don’t burn any tokens

BurnDestination

ByteReadWeight

CandidacyBond

CandidacyDelay

Require 3 days in blocks for each candidate to be allowed for the election.

CooloffPeriod

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DataDepositPerByte

DealWithFees

Deals with transaction fees

DefaultElasticity

DepositLockStrategy

DesiredMembers

DesiredRunnersUp

DisabledValidatorsThreshold

DockChainId

Keeping 22 as its the ss58 prefix of mainnet

ElectionsPhragmenPalletId

EnactmentPeriod

EpochDuration

ExistentialDeposit

ExpectedBlockTime

FastTrackVotingPeriod

FieldDeposit

FindAuthorTruncated

GasPrice

Fixed gas price. Gas price is always 50 mirco-token (0.00005 token) per gas.

GasWeightMap

GenesisConfig

HighRateRewardDecayPct

ImOnlineUnsignedPriority

InstantAllowed

LaunchPeriod

LowRateRewardDecayPct

MaxAdditionalFields

MaxApprovals

MaxCandidates

MaxElectableTargets

MaxElectingVoters

MaxNominations

MaxNominatorRewardedPerValidator

A validator will only have 256 nominators, so max (rewarded) nominators is 256*50 = 12800. This number is smaller than current token holders (20K) but not all holder participate. Keeping it small to batch payout extrinsics

MaxProposals

MaxRegistrars

MaxScheduledPerBlock

MaxSubAccounts

MaxSwapBonuses

MaxSymbolBytesLen

MaxVestingBonuses

MaxVoters

MaxVotes

MaximumBlockWeight

We allow for 1 seconds of compute with a 3 second average block time.

MaximumReasonLength

MinerMaxWeight

MinimumDeposit

1000 tokens

MinimumMultiplier

MinimumPeriod

MultiPhaseUnsignedPriority

NoPreimagePostponement

NposSolution16

OffchainRepeat

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution weight.

OffencesWeightSoftLimit

OffendingValidatorsThreshold

OnChainSeqPhragmen

Origin

The runtime origin type representing the origin of a call.

PSPublicKey

Public key for the PS signature scheme.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PostUpgradeHighRateDuration

PrecompilesValue

PreimageByteDeposit

0.02 token

ProposalBond

ProposalBondMinimum

ReportLongevity

Specifies the number of blocks for which the equivocation is valid.

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SessionsPerEra

ShortTimeProposal

Proposal with lifetime less than 2 hours (in blocks) requires to be approved by all members.

SignedDepositBase

SignedDepositByte

SignedPhase

SignedRewardBase

SlashDeferDuration

SpendPeriod

StakingBenchmarkingConfig

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

SubAccountDeposit

TargetBlockFullness

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

.0001 token

TransactionConverter

TreasuryPalletId

Matches treasury account created during PoA.

TreasuryRewardsPct

U64CurrencyToVote

UncleGenerations

UnsignedPhase

Version

VestingDuration

Vesting duration in number of blocks. Duration is 183 days and block time is 3 sec. (183 * 24 * 3600) / 3 = 5270400

VestingMilestones

Number of vesting milestones

VotingBondBase

VotingBondFactor

VotingPeriod

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

EPMCall

Contains one variant per dispatchable that can be called by an extrinsic.

Event

OffchainPublicKey

Public key for different signature schemes. Currently, can be either BBS, BBS+, Pointcheval-Sanders or BBDT16.

OriginCaller

StakerStatus

Indicates the initial status of the staker.

SystemCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 10% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 10%.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

DAY

DOCK

The token has 6 decimal places

HOUR

MAXIMUM_BLOCK_WEIGHT

We allow for 1 second of compute with a 3 second average block time.

MILLISECONDS_PER_YEAR

Era duration should be less than or equal year

MILLISECS_PER_BLOCK

MINUTE

PRIMARY_PROBABILITY

SLOT_DURATION

VERSION

This runtime version. Whe compiling with feature “mainnet”, spec_name will be “dock-main-runtime”, when compiling with feature “testnet”, spec_name will be “dock-test-runtime”, when not using either of those, spec_name will be “dock-dev-runtime”

WASM_BINARY

WEEK

WEIGHT_PER_GAS

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Accumulator

Address

The address format for describing accounts.

Agreement

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Anchor

Attest

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balance

Balance of an account.

Balances

BalancesConfig

BaseFee

BaseFeeConfig

BlobStore

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

Bounties

CheckedExtrinsic

Extrinsic type that has already been checked.

CheqdMigration

Council

CouncilConfig

DIDModule

DIDModuleConfig

Democracy

EVM

EVMConfig

ElectionProviderMultiPhase

Elections

ElectionsConfig

Ethereum

EthereumConfig

GrandpaFinality

GrandpaFinalityConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Index

Index of a transaction in the chain.

Master

MasterConfig

MigrationModule

Moment

Type used for expressing timestamp.

OffchainSignatures

Offences

PoAModule

PoAModuleConfig

PriceFeedModule

Revoke

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Staking

StakingConfig

StakingRewards

StatusListCredential

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeMembership

TechnicalCommitteeMembershipConfig

Timestamp

Tips

TransactionPayment

Treasury

TreasuryConfig

TrustRegistry

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility