Crate dock_runtime

Dock testnet runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use poa;

pub use price_feed;

pub use simple_democracy;

pub use token_migration;

Modules

anchor
Minimal proof of existence registry.

api

attest
This module allows DIDs to publically attests to arbirary (and arbitrarily large) RDF claimgraphs. These attestations are not stored on-chain; rather, the attester chooses a storage method by specifying an Iri.

blob
Generic immutable single-owner storage.

did

master
Simulates a multisig root account.

opaque
Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

revoke

weight_to_fee

Structs

BaseFilter

BlockGasLimit

BlockHashCount

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DockChainId

EnactmentPeriod

EthereumFindAuthor

ExistentialDeposit

FastTrackVotingPeriod

GenesisConfig

InstantAllowed

LaunchPeriod

MaxBlobSize

MaxLocks

MaxProposals

MaxScheduledPerBlock

MaxVotes

MaximumBlockWeight
We allow for 1 seconds of compute with a 3 second average block time.

MinimumDeposit
1000 tokens

MinimumPeriod

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageByteDeposit
0.1 token

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights
After each block we

SS58Prefix

StorageWeight

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TransactionByteFee
.01 token

TransactionConverter

UncleGenerations
Not accepting any uncles

UnitGasPrice
Fixed gas price of 1.

Version

VestingDuration
Vesting duration in number of blocks. Duration is 183 days and block time is 3 sec. (183 * 24 * 3600) / 3 = 5270400

VestingMilestones
Number of vesting milestones

VotingPeriod

Enums

Call

Event

OriginCaller

Constants

AVERAGE_ON_INITIALIZE_RATIO
We assume that an on-initialize consumes 10% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 10%.

DAYS

DOCK
The token has 6 decimal places

HOURS

MAXIMUM_BLOCK_WEIGHT
We allow for 1 second of compute with a 3 second average block time.

MILLISECS_PER_BLOCK

MINUTES

VERSION
This runtime version.

WASM_BINARY

Functions

native_version
The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address
The address format for describing accounts.

Anchor

Attest

Aura

AuraConfig

Authorship

Balance
Balance of an account.

Balances

BalancesConfig

BlobStore

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

BlockNumber
An index to a block.

CheckedExtrinsic
Extrinsic type that has already been checked.

Council

CouncilConfig

CouncilMembership

CouncilMembershipConfig

DIDModule

DIDModuleConfig

Democracy

EVM

EVMConfig

Ethereum

EthereumConfig

Grandpa

GrandpaConfig

Hash
A hash of some data used by the chain.

Index
Index of a transaction in the chain.

Master

MasterConfig

MigrationModule

PoAModule

PoAModuleConfig

PriceFeedModule

PriceFeedModuleConfig

RandomnessCollectiveFlip

Revoke

Scheduler

Session

SessionConfig

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock
A Block signed with a Justification

SimpleDemocracy

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeMembership

TechnicalCommitteeMembershipConfig

Timestamp

TransactionPayment

Utility